+++ src/window_auto_adjust_size.c
@@ -31,8 +31,8 @@ static ret_t on_change_to_chinese(void *ctx, event_t *e) {
   return RET_OK;
 }
 
-ret_t window_auto_adjust_size_open(void) {
-  widget_t *win = window_open("auto_adjust_size");
+ret_t window_auto_adjust_size_open(const char* name) {
+  widget_t *win = window_open(name);
   widget_child_on(win, "english", EVT_CLICK, on_change_to_english, NULL);
   widget_child_on(win, "chinese", EVT_CLICK, on_change_to_chinese, NULL);