+++ src/window_auto_adjust_size.h
@@ -26,7 +26,7 @@
 
 BEGIN_C_DECLS
 
-ret_t window_auto_adjust_size_open(void);
+ret_t window_auto_adjust_size_open(const char* name);
 
 END_C_DECLS