+++ docs/changes.md
@@ -1,5 +1,8 @@
 # 最新动态
 
+2020/04/28
+ * 增加label的 margin left, right, top, bottom不同时的测试用例。
+
 2020/04/27
  * 增加auto ajust size时指定max_w参数的测试用例。